telemetry: share RPC conventions instead of importing bsync from bsky
The AppView labelled its bsync client metrics by importing constants from
@atproto/bsync, so that both halves of a call would agree on the dimensions.
That made a service package a runtime dependency of another service package,
which is a heavy way to share five string constants and a switch.

Move them to @atproto-labs/opentelemetry-node, which both already depend on:
the bsync.namespace/bsync.operation attribute keys and
RPC_CALL_DURATION_BUCKETS to the /conventions entrypoint, and
statusCodeToString() to a new /util entrypoint. The attribute keys were
previously bare string literals duplicated across three call sites with a
comment asking future readers to keep them in sync; they are now constants.
