+++ packages/bsky/src/bsync.ts
@@ -9,6 +9,10 @@ import {
   type ConnectTransportOptions,
   createConnectTransport,
 } from '@connectrpc/connect-node'
+import {
+  ATTR_BSYNC_NAMESPACE,
+  ATTR_BSYNC_OPERATION,
+} from '@atproto-labs/opentelemetry-node/conventions'
 import { Service } from './proto/bsync_connect.js'
 import { Method, type PutOperationRequest } from './proto/bsync_pb.js'
 import { createRpcClientInterceptor } from './telemetry/rpc.js'
@@ -21,15 +25,12 @@ export const createBsyncClient = (
   const transport = createConnectTransport({
     ...opts,
     interceptors: [
-      // @NOTE These attribute keys must match the ones bsync sets on its own
-      // handler metrics (see `withRpcServerTelemetry` in @atproto/bsync), so
-      // that both sides of a call can be plotted on the same dimensions.
       createRpcClientInterceptor((req) => {
         if (req.method.name !== Service.methods.putOperation.name) return {}
         const { namespace, method } = req.message as PutOperationRequest
         return {
-          'bsync.namespace': namespace,
-          'bsync.operation': Method[method]?.toLowerCase() ?? 'unknown',
+          [ATTR_BSYNC_NAMESPACE]: namespace,
+          [ATTR_BSYNC_OPERATION]: Method[method]?.toLowerCase() ?? 'unknown',
         }
       }),
       ...(opts.interceptors ?? []),
