+++ packages/bsky/tsconfig.build.json
@@ -4,7 +4,6 @@
   "exclude": ["./src/**/*.test.ts"],
   "references": [
     { "path": "../api/tsconfig.build.json" },
-    { "path": "../bsync/tsconfig.build.json" },
     { "path": "../common/tsconfig.build.json" },
     { "path": "../crypto/tsconfig.build.json" },
     { "path": "../did/tsconfig.build.json" },
