+++ packages/bsync/src/telemetry/attributes.ts
@@ -1,12 +1,13 @@
 import type { Attributes } from '@opentelemetry/api'
+import {
+  ATTR_BSYNC_NAMESPACE,
+  ATTR_BSYNC_OPERATION,
+} from '@atproto-labs/opentelemetry-node/conventions'
 import { Method } from '../proto/bsync_pb.js'
 
 /**
  * Attributes describing a stash operation.
  *
- * @note These attribute keys must match the ones the AppView sets on its own
- * client metrics (see `createBsyncClient` in @atproto/bsky), so that both sides
- * of a call can be plotted on the same dimensions.
  *
  * Both values are low-cardinality: the namespace is an NSID drawn from a fixed
  * set of record types, and the method is one of create/update/delete.
@@ -15,6 +16,6 @@ export const operationAttributes = (op: {
   namespace: string
   method: Method
 }): Attributes => ({
-  'bsync.namespace': op.namespace,
-  'bsync.operation': Method[op.method]?.toLowerCase() ?? 'unknown',
+  [ATTR_BSYNC_NAMESPACE]: op.namespace,
+  [ATTR_BSYNC_OPERATION]: Method[op.method]?.toLowerCase() ?? 'unknown',
 })
