+++ services/bsky/Dockerfile
@@ -12,12 +12,14 @@ COPY ./pnpm-workspace.yaml ./
 COPY ./tsconfig ./tsconfig
 COPY ./lexicons ./lexicons
 
+# bsync is not a runtime dep but bsky's "codegen:buf" script reads ../bsync/proto at build time
+COPY ./packages/bsync/proto ./packages/bsync/proto
+
 # NOTE bsky's transitive dependencies go here: if that changes, this needs to be updated.
 # pnpm list --filter 'bsky-app-view-service...' --depth 0 --json | jq '.[].path' | sort -u | sed "s|\"$PWD/|./|; s|\"\$||"
 COPY ./packages/api ./packages/api
 COPY ./packages/aws ./packages/aws
 COPY ./packages/bsky ./packages/bsky
-COPY ./packages/bsync ./packages/bsync
 COPY ./packages/common-web ./packages/common-web
 COPY ./packages/common ./packages/common
 COPY ./packages/crypto ./packages/crypto
